% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aFIPC.R
\name{autoFIPC}
\alias{autoFIPC}
\title{automated fixed item parameter linking}
\usage{
autoFIPC(
  newformXData,
  oldformYData,
  newformCommonItemNames,
  oldformCommonItemNames,
  itemtype = "3PL",
  newformBILOGprior = NULL,
  oldformBILOGprior = NULL,
  tryFitwholeNewItems = T,
  tryFitwholeOldItems = T,
  checkIPD = T,
  tryEM = T,
  freeMEAN = T,
  forceNormalZeroOne = F,
  parameterOverwrite = F,
  empiricalhist = F,
  confirmCommonItems = NULL,
  ...
)
}
\arguments{
\item{newformXData}{new form data X}

\item{oldformYData}{old form (base form) data Y}

\item{newformCommonItemNames}{Common item variable names in new form data}

\item{oldformCommonItemNames}{Common item variable names in old (base) form data}

\item{itemtype}{itemtype of calibration}

\item{newformBILOGprior}{using BILOG-MG prior when try to calibrate 3PL model? if you want, set the this to TRUE}

\item{oldformBILOGprior}{using BILOG-MG prior when try to calibrate 3PL model? if you want, set the this to TRUE}

\item{tryFitwholeNewItems}{do you want to try calibrate new form data? default is TRUE}

\item{tryFitwholeOldItems}{do you want to try calibrate base form data? default is TRUE}

\item{checkIPD}{do you want to check item parameter drift? default is TRUE}

\item{tryEM}{do you want to try EM algorithm when you calibrate model? defalut is TRUE}

\item{freeMEAN}{allow freely mean estimation, default is TRUE}

\item{forceNormalZeroOne}{set the prior distribution follows N(0,1) distribution. default is FALSE}

\item{parameterOverwrite}{don't touch it}

\item{empiricalhist}{do you want to use empirical histogram method when tryEM = TRUE? default is FALSE}

\item{confirmCommonItems}{set TRUE to accept the supplied common-item pairs without an interactive prompt. Default NULL: in interactive sessions the user will be prompted; set FALSE to explicitly reject (function will stop).}

\item{...}{Additional arguments reserved for future extensions.}
}
\value{
the model list of the base form, new form, linked form
}
\description{
automated fixed item parameter linking
}
\examples{
\dontrun{
autoFIPC(
  newformXData = new_model,
  oldformYData = old_model,
  newformCommonItemNames = common_new,
  oldformCommonItemNames = common_old,
  confirmCommonItems = TRUE
)
}
}
